requirements.txt
# requirements.txt

Django>=5.2.0,<5.3.0
Pillow>=10.0.0
whitenoise>=6.5.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
djangorestframework>=3.14.0
python-decouple>=3.8
requests>=2.31.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.7
dj-database-url>=2.1.0
django-debug-toolbar>=4.2.0
django-image-cropping>=1.7.0
easy-thumbnails>=2.8.5
cloudinary==1.36.0
django-cloudinary-storage==0.3.0